requirements.txt
Django==5.0.7
djangorestframework==3.15.2
channels==4.1.0
channels-redis==4.2.0
celery==5.4.0
redis==5.0.8
psycopg2-binary==2.9.9
requests==2.32.3
dj-database-url==2.2.0
gunicorn==22.0.0
uvicorn==0.30.1
whitenoise==6.7.0
pytest==8.3.2
pytest-django==4.8.0
pytest-asyncio==0.23.8
